fix: Header type fixes, struct padding, and symbol alignment
- Fix bool32_t/bool return type mismatches on 25+ virtual methods
- Fix const qualifier mismatches on 16 non-virtual methods
- Fix void→uint32_t return type on 14 methods
- Fix struct/class forward declarations for symbol matching
- Fix reference-vs-pointer parameter mismatches (MapCoords)
- Fix uint32_t → unsigned long (K/I mangling) across 48 headers
- Add padding fields to 76 structs for correct sizeof in destructors
- Change 49 bool32_t fields to bool in C++ class definitions
- Restructure GContainerInfo hierarchy and fix GTreeInfo size
- Add Base::operator delete for correct vtable layout

+++ black/LocalBase.h
@@ -29,7 +29,7 @@ public:
     // Static methods
 
     // win1.41 005f8790 mac 1009eb9c LocalBase::operator new(unsigned long)
-    static LocalBase* __nw(size_t size, const char* file_name, uint32_t line);
+    static LocalBase* __nw(unsigned long size, const char* file_name, uint32_t line);
 
     // Constructors
