+++ black/MPFEChooseConnection.h
@@ -32,7 +32,7 @@ public:
     // Non-virtual methods
 
     // win1.41 0062b030 mac 1039fa00 MPFEChooseConnection::Init(unsigned long, unsigned long, void (*)(int, SetupBox *, SetupControl *, int, int))
-    void Init(uint32_t param_1, uint32_t param_2, void (* callback)(int, SetupBox *, SetupControl *, int, int));
+    void Init(unsigned long param_1, unsigned long param_2, void (* callback)(int, SetupBox *, SetupControl *, int, int));
 };
 
 #else // __cplusplus
